add interface implementation to services
