+++ src/main/java/com/invoicemang/invoicemangbackendapi/controller/PurchaseController.java
@@ -1,7 +1,7 @@
 package com.invoicemang.invoicemangbackendapi.controller;
 
 import com.invoicemang.invoicemangbackendapi.model.Purchase;
-import com.invoicemang.invoicemangbackendapi.service.PurchaseService;
+import com.invoicemang.invoicemangbackendapi.service.PurchaseServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -15,15 +15,11 @@ import java.util.List;
 @CrossOrigin(origins = "*")
 public class PurchaseController {
 
-    private PurchaseService purchaseService;
-
     @Autowired
-    public PurchaseController(PurchaseService purchaseService) {
-        this.purchaseService = purchaseService;
-    }
+    private PurchaseServiceImpl purchaseServiceImpl;
 
     @GetMapping
     public List<Purchase> getPurchases() {
-        return purchaseService.getPurchases();
+        return purchaseServiceImpl.getPurchases();
     }
 }
